code added for Transaction history to work

+++ src/application/Model.java
@@ -15,16 +15,16 @@ public class Model {
 		public float calculate(long number1, long number2, String operator){
 			switch(operator){
 			case "+":
-				return number1 + number2;
+				return (float)number1 + (float)number2;
 			case "-":
-				return number1 - number2;
+				return (float)number1 - (float)number2;
 			case "/":
 				if(number2 == 0)
 					return 0;
 				else
-					return number1 / number2;
+					return (float)number1 / (float)number2;
 			case "*":
-				return number1 * number2;
+				return (float)number1 * (float)number2;
 			default:
 				break;
 			}
